requirements.txt
Django==3.1.3
PyNaCl==1.4.0
